+++ lib/pitel_sdk/pitel_call.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import 'package:flutter_easyloading/flutter_easyloading.dart';
 import 'package:flutter_webrtc/flutter_webrtc.dart';
 import 'package:flutter_pitel_voip/component/pitel_call_state.dart';
 import 'package:flutter_pitel_voip/component/pitel_rtc_video_renderer.dart';
@@ -6,8 +7,12 @@ import 'package:flutter_pitel_voip/component/pitel_ua_helper.dart';
 import 'package:flutter_pitel_voip/component/sip_pitel_helper_listener.dart';
 import 'package:flutter_pitel_voip/pitel_sdk/pitel_log.dart';
 import 'package:flutter_pitel_voip/sip/sip_ua.dart';
+import 'package:throttling/throttling.dart';
+import 'package:connectivity_plus/connectivity_plus.dart';
 
 import 'pitel_client.dart';
+
+final thr = Throttling(duration: const Duration(milliseconds: 2000));
 
 class PitelCall implements SipUaHelperListener {
   final PitelLog _logger = PitelLog(tag: 'PitelCall');
@@ -41,6 +46,14 @@ class PitelCall implements SipUaHelperListener {
   bool get isHaveCall => _callIdCurrent?.isNotEmpty ?? false;
   String? _callIdCurrent;
   bool isBusy = false;
+  String _outPhone = "";
+  ConnectivityResult _checkConnectivity = ConnectivityResult.none;
+
+  String get outPhone => _outPhone;
+
+  void resetOutPhone() {
+    _outPhone = "";
+  }
 
   void setCallCurrent(String? id) {
     _callIdCurrent = id;
@@ -443,30 +456,43 @@ class PitelCall implements SipUaHelperListener {
     isBusy = true;
   }
 
-  Future<String> outgoingCall(
-    String phoneNumber,
-    VoidCallback onRegister,
-  ) async {
-    try {
-      final isRegistered = getRegisterState();
-      if (isRegistered == 'Registered') {
-        final res = await _handleCall(phoneNumber);
-        return res;
-      } else {
-        onRegister();
-        await Future.delayed(const Duration(milliseconds: 500));
-        final res = await _handleCall(phoneNumber);
-        return res;
-      }
-    } catch (error) {
-      return error.toString();
-    }
-  }
+  void outGoingCall({
+    required String phoneNumber,
+    required VoidCallback handleRegisterCall,
+  }) {
+    thr.throttle(() async {
+      _outPhone = phoneNumber;
+      final PitelCall pitelCall = PitelClient.getInstance().pitelCall;
+      final PitelClient pitelClient = PitelClient.getInstance();
 
-  Future<String> _handleCall(String phoneNumber) async {
-    final PitelClient pitelClient = PitelClient.getInstance();
-    return pitelClient
-        .call(phoneNumber, true)
-        .then((value) => value.fold((succ) => "OK", (err) => err.toString()));
+      final connectivityResult = await (Connectivity().checkConnectivity());
+      if (connectivityResult == ConnectivityResult.none) {
+        EasyLoading.showToast(
+          'Please check your network',
+          toastPosition: EasyLoadingToastPosition.center,
+        );
+        return;
+      }
+      if (connectivityResult != _checkConnectivity) {
+        _checkConnectivity = connectivityResult;
+        EasyLoading.show(status: "Connecting...");
+        handleRegisterCall();
+        return;
+      }
+      final isRegistered = pitelCall.getRegisterState();
+      if (isRegistered == 'Registered') {
+        pitelClient
+            .call(phoneNumber, true)
+            .then((value) => value.fold((succ) => "OK", (err) {
+                  EasyLoading.showToast(
+                    err.toString(),
+                    toastPosition: EasyLoadingToastPosition.center,
+                  );
+                }));
+      } else {
+        EasyLoading.show(status: "Connecting...");
+        handleRegisterCall();
+      }
+    });
   }
 }